got the park to render to the DOM once it is selected
