+++ index.html
@@ -33,7 +33,7 @@
 
         <div id="national-parks"class="plans">
             <div class="weather"></div>
-            <button class="details-button">DETAILS</button>
+            <!-- <button class="details-button">DETAILS</button> -->
             <p class="details"></p>
         </div>
