+++ scripts/main.js
@@ -1,5 +1,5 @@
 import { parkSelect } from "./parks/ParkSelect.js";
-import "./parks/ParkList.js"
+import "./parks/ParkPreview.js"
