+++ scripts/parks/Park.js
@@ -1,7 +1,11 @@
 export const Park = (parkObj) => {
     return `
-    <p class="park">
-        
+    <p id="national-parks">
+        ${parkObj.fullName}
+        <button class="park-detail">Details</button>
         </p>
         `
 }
+
+// console.log("Full Name", parksObj.fullName)
+
